try to run on ubuntu

+++ proj1/proj1/Raytracer.h
@@ -14,16 +14,16 @@ private:
 	int scale;
 	int imgHeight;
 	int imgWidth;
-	float MAX_LIGHT_LEN = 1000000;
+	double MAX_LIGHT_LEN = 1000000;
 	int MAX_DEPTH = 10;
 public:
 	RayTracer(Camera* c, Scene* s, int height, int width,int sc) :camera(c), scene(s), imgHeight(height), imgWidth(width), scale(sc){}
 	cv::Mat getImg() { return result; }
 	void showImg();
 	void writeImg();
-	float get_shadow(Object* obj, Vector3 intersectPos, int i, Vector3& L);
+	double get_shadow(Object* obj, Vector3 intersectPos, int i, Vector3& L);
 
-	Color trace(Ray* r, int depth, float length, float refract_idx, float &inter_l, int &inter_id);
+	Color trace(Ray* r, int depth, double length, double refract_idx, double &inter_l, int &inter_id);
 	void run();
 };
