+++ proj1/proj1/common.h
@@ -9,8 +9,8 @@ typedef Vector3 Color;
 
 class Point {
 public:
-	float x, y;
-	Point(float _x,float _y):x(_x),y(_y){}
+	double x, y;
+	Point(double _x,double _y):x(_x),y(_y){}
 };
 
 class Ray{
@@ -37,11 +37,11 @@ class Light {
 public:
 	Vector3 o;
 	Color color;
-	float radius;
-	float brightness;
+	double radius;
+	double brightness;
 	Light():o(Vector3(0,0,0)),color(Color(0,0,0)),radius(10000),brightness(0) {}
-	Light(Vector3 origin, Color c, float r,float b) : o(origin),color(c),radius(r),brightness(b){}
-	void set(Vector3 origin, Color c, float r,float b) {
+	Light(Vector3 origin, Color c, double r,double b) : o(origin),color(c),radius(r),brightness(b){}
+	void set(Vector3 origin, Color c, double r,double b) {
 		o = origin;
 		color = c;
 		radius = r;
