+++ proj1/proj1/scene.h
@@ -4,12 +4,12 @@
 
 struct IntersectPoint {
 	Object* obj;
-	float t;
+	double t;
 	int obj_id;
 	bool inside;
 	IntersectPoint(){}
-	IntersectPoint(Object* o,float _t,bool in,int id):obj(o),t(_t),inside(in),obj_id(id){}
-	void set(Object* o, float _t, bool in, int id) { obj = o; t = _t; inside = in; id = obj_id; }
+	IntersectPoint(Object* o,double _t,bool in,int id):obj(o),t(_t),inside(in),obj_id(id){}
+	void set(Object* o, double _t, bool in, int id) { obj = o; t = _t; inside = in; id = obj_id; }
 };
 
 class Scene {
@@ -21,11 +21,11 @@ private:
 public:
 	Scene();
 
-	Meterial* getMeterial(float reflect, float refract,float diffuse, float specular);
+	Meterial* getMeterial(double reflect, double refract,double diffuse, double specular);
 	ColorTexture* getColorTexture(Color c);
 	PicTexture* getPicTexture(std::string path);
-	Sphere* getSphere(Meterial* m, Texture* t, Vector3 o, float r);
-	Plane* getPlane(Meterial* m, Texture* t, Vector3 P, Vector3 n, float D);
+	Sphere* getSphere(Meterial* m, Texture* t, Vector3 o, double r);
+	Plane* getPlane(Meterial* m, Texture* t, Vector3 P, Vector3 n, double D);
 
 	void addObj(Object* obj);
 	int getObjCnt();
